requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
torch
gradio
